requirements.txt
annotated-types==0.7.0
anyio==4.8.0
apify_client==1.8.1
apify_shared==1.1.2
certifi==2025.1.31
click==8.1.8
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kafka-python==2.0.2
loguru==0.7.2
more-itertools==10.6.0
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2025.1
schedule==1.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.28.1
